Revert "Library updates"
This reverts commit c45f17049d9996e18cc561bd0f3524f3aef44108.

+++ JuceLibraryCode/modules/juce_core/native/juce_android_Network.cpp
@@ -260,7 +260,7 @@ static void findIPAddresses (int sock, Array<IPAddress>& result)
     do
     {
         bufferSize *= 2;
-        buffer.calloc (bufferSize);
+        buffer.calloc ((size_t) bufferSize);
 
         cfg.ifc_len = bufferSize;
         cfg.ifc_buf = buffer;
